% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Plot FARS Data By State}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{An integer corresponding to the state number. Valid values are 1 to 51, except 3}

\item{year}{A year provided as 4 digit integer or string}
}
\value{
This function plots a heat map of the accident for the given state and year.
}
\description{
Function filter data for given state and use it to plot a heat map where area with more accidents
will be shown in darker spots. The function will print a message "no accident to plot"
if there is no accident to report.
}
\note{
Though the function anticipates giving a message if no accident is reported
for a state in a given year, there is a very good chance that this portion of the
code will not be trigerred due to a check to ensure the \code{state.num} filter
parameter is in the data set.
}
\examples{
   \dontrun{fars_map_state(1,2015)}

}
