% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{Read FARS Data}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{The name of the file to be read to tbl_df object}
}
\value{
Tibble with data from the supplied file.
}
\description{
The function load data from the US National Highway Traffic Safety Administration's Fatality
Analysis Reporting System (FARS):
\url{https://www.nhtsa.gov/research-data/fatality-analysis-reporting-system-fars}
}
\examples{
   \dontrun{fars_read("accident_2014.csv.bz2")}

}
