% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Read multiple years of FARS Data}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{Vector of years denoted as 4 digit integers or strings}
}
\value{
Function returns a list of data frames with FARS data
}
\description{
Funtion takes vector of years and converted each year to FARS filename.
Each data frame contains just the year and month. If FARS data file
is not available for any year, the list entry for that year is NULL
}
\examples{
   \dontrun{fars_read_years(c(2013,2015))}
   \dontrun{fars_read_years(2013:2015)}

}
