% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Summarize multiple years FARS Data}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{Vector of years denoted as 4 digit integers or strings}
}
\value{
Returns data frame with FARS data summarized by years and months.
}
\description{
Multiple years FARS data is summarized
by counting the number of rows for each year and month.
The summarized data frame displays months as rows and years as columns.
}
\examples{
   \dontrun{fars_summarize_years(c(2013,2015))}
   \dontrun{fars_summarize_years((2013:2015))}

}
