% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{Construct filename with provided year}
\usage{
make_filename(year)
}
\arguments{
\item{year}{A year provided as 4 digit integer or string}
}
\value{
FARS filename for given year
}
\description{
Function expects \code{year} as input and
constructs the FARS filename for the given year.
}
\examples{
   \dontrun{make_filename(2014)}
   \dontrun{make_filename("2014")}

}
