% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{fars_map_state creates a map of fatalities in a state}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{A state number}

\item{year}{One of 2013, 2014, or 2015}
}
\value{
a map of the fatalities in the state represented by state.num
}
\description{
fars_map_state creates a map of fatalities in a state
}
\examples{
require(maps)
fars_map_state(48, 2013)
fars_map_state(6, 2015)
}

