% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_util.R
\name{fars_read}
\alias{fars_read}
\title{Reads a CSV file as a tibble (tbl_df).}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{string path to a CSV file to be read.
The path may be absolute or relative to the current working directory.}
}
\value{
tbl_df representation of the data in the input CSV file.
}
\description{
Reads a CSV file as a tibble (tbl_df).
}
\details{
Data is from the US National Highway Traffic Safety
    Administration's Fatality Analysis Reporting System
    (\href{https://www.nhtsa.gov/research-data/fatality-analysis-reporting-system-fars}{FARS}).
}
\note{
Fails when there is no file with the given 'filename'.
}
\examples{
    # my_tbl <- fars_read("inst/extdata/fars_data/accident_2013.csv.bz2")
}
