% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{(Internal use only) fars_read reads in a FARS data file}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{A string which represents a file name}
}
\value{
a data.frame
}
\description{
(Internal use only) fars_read reads in a FARS data file
}
\note{
If the file does not exist the function exits
}

