% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_util.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Reads a list of FARS files for the given years.}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{vector of integer or string years for data to be read.}
}
\value{
list of tibbles with columns MONTH and year, corresponding
    to the original rows in the data files.
    When there is no file for a given year, that entry will
    be NULL and a warning is given.
}
\description{
The list returned by this function is useful for
    summarizing numbers of accidents by month for certain years.
}
\examples{
    fars_read_years(c(2013, 2014))
}
