% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{(Internal use only) fars_read_years extracts data by year}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{A year or a list of years in the range 2013 to 2015}
}
\value{
This function returns a list of tibbles, where each tibble
  contains the year and month from the observations in the corresponding
  year's FARS data. If the year is not valid, the list will be NULL.
}
\description{
(Internal use only) fars_read_years extracts data by year
}
\note{
If the year is invalid the function exits with a warning
}

