% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_util.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Summarizes the number of operations by month and year.}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{vector of integer or string years for data to be read.}
}
\value{
list of tibbles with columns MONTH and year.
}
\description{
Summarizes the number of operations by month and year.
}
\note{
If there is no data for one of the input years, a warning is
    given and that year will be ignored.
}
\examples{
    by_month <- fars_summarize_years(c(2013, 2014))
}
