% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{fars_summarize_years creates a summary from FARS data files. For
    each year given, a monthly total of fatalities is returned.}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{A year or a list of years from 2013 to 2015 inclusive.}
}
\value{
summary of fatalities by month and year, a 12 x number of years
    tibble.
}
\description{
fars_summarize_years creates a summary from FARS data files. For
    each year given, a monthly total of fatalities is returned.
}
\examples{
fars_summarize_years(2015)
fars_summarize_years(c(2013,2014,2015))
}

