% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_util.R
\name{make_filename}
\alias{make_filename}
\title{Builds a string filename for a
    \href{https://www.nhtsa.gov/research-data/fatality-analysis-reporting-system-fars}{FARS}
    file for the given year.}
\usage{
make_filename(year)
}
\arguments{
\item{year}{of data to be read. May be string or integer.}
}
\value{
string filename, following the naming convention of
    \href{https://www.nhtsa.gov/research-data/fatality-analysis-reporting-system-fars}{FARS}
    files.
}
\description{
Builds a string filename for a
    \href{https://www.nhtsa.gov/research-data/fatality-analysis-reporting-system-fars}{FARS}
    file for the given year.
}
\examples{
    fars_2013_path <- make_filename("2013")
}
