% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{(Internal use only) make_filename Creates a filename with the correct structure for FARS data}
\usage{
make_filename(year)
}
\arguments{
\item{year}{A 4-digit year, e.g., 2015}
}
\value{
a string which represents a FARS data file name
}
\description{
(Internal use only) make_filename Creates a filename with the correct structure for FARS data
}

